esa_dropshipping_fastapi_mvp/esa_dropshipping/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.9.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
